# reorder to move parts into functions, switch to c++20 and gcc