- Moved the various parts into functions to show that one doesn't have to use a single large main function.
- some small performance improvements int top5 logic by improving memory locality.
- moved to gcc & c++20 since that was the fastest after changes.

@jinyus would be curious to see how this performs on the VM.

(Tried adding better maps but only got better results when using `march=x86_64_v3` but seems like that is forbidden here from what I understand)