@jinyus 2023-11-01
Main:
`C++ | 24.92 ms | 339.00 ms | 2.92 s`

This Pr:
```
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 23 ms
        Processing time (w/o IO): 23 ms
        Processing time (w/o IO): 23 ms
        Processing time (w/o IO): 23 ms
        Processing time (w/o IO): 23 ms
        Processing time (w/o IO): 23 ms
        Processing time (w/o IO): 23 ms
        Processing time (w/o IO): 23 ms
        Processing time (w/o IO): 24 ms
        Processing time (w/o IO): 23 ms
        Processing time (w/o IO): 23 ms
        Processing time (w/o IO): 23 ms
        Processing time (w/o IO): 23 ms
          Time (mean ± σ):     226.8 ms ±   2.4 ms    [User: 194.4 ms, System: 31.6 ms]
          Range (min … max):   222.4 ms … 230.1 ms    10 runs
         
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 328 ms
        Processing time (w/o IO): 323 ms
        Processing time (w/o IO): 322 ms
          Time (mean ± σ):      1.148 s ±  0.002 s    [User: 1.008 s, System: 0.133 s]
          Range (min … max):    1.146 s …  1.150 s    2 runs
         
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 2789 ms
        Processing time (w/o IO): 2781 ms
        Processing time (w/o IO): 2795 ms
          Time (mean ± σ):      5.415 s ±  0.112 s    [User: 4.940 s, System: 0.334 s]
          Range (min … max):    5.336 s …  5.493 s    2 runs

```